TrainingRecapTable refactoring
